Location IDs as integers, not strings.

+++ AlbionDataAvalonia/Network/Models/MarketHistoriesUpload.cs
@@ -6,7 +6,7 @@ public class MarketHistoriesUpload
 {
     public uint AlbionId;
     public string AlbionIdString;
-    public ushort LocationId;
+    public int LocationId;
     public byte QualityLevel;
     public Timescale Timescale;
     public List<MarketHistory> MarketHistories = new List<MarketHistory>();
